Fix timeouts on testbarrierconf with exact CPU commit
In spin-polling mode, the performance of this test is particularly
sensitive to system noise on any of its expected cores. This was
causing timeouts on at least the WSL scheduler when the core count
exactly matches the test thread count.

Add TEST_SET_WAITMODE_RESERVE to allow such tests to reserve extra cores
to accomodate system noise.

Apply the new utility in both SEQ and PAR modes of the test, as the
SEQ mode demonstrates a similar problem (albeit without polling threads).