ibv: improve specialization of AM injection
This commit adds a boolean argument to gasnetc_ReqRepGeneric() to
distinguish Requests from Replies.  Technically this is redundant
since it is exactly equivalent to the "NULLness" of `token`.  However,
compilers (clang, gcc and xlc) are producing better code with the
added argument, presumably because the `token` argument in a Reply
cannot be proven to be non-NULL.  Addition of `gasneti_assume(token)`
in the Reply callers was not sufficient.