acinclude: identify Clang compiler sub-family
This commit expands the handling of "*_SUBFAMILY" to include
sub-families for clang.  The default is 'LLVM' (analogous to 'GNU' for
gcc).  The two new sub-families are 'APPLE' and 'CRAY' (the latter for
"CCE 9.0.0").  Use of all-caps is consistent with existing 'APPLE' and
'NVIDIA' sub-families for gcc.

Additionally, the sub-family is now included in configure's output.

While there is not yet any consumer for this information, both vendors
are known to make non-trivial changes relative to the baseline,
including the addition of additional command line options (which one
could reasonably want to "reason about" at configure time).