fixup: aries/gemini: remove all gemini-specific...
This commit removes lingering code to allocate, use, and destroy
`destination_cq_handle`, which was specific to the `FIX_HT_ORDERING`
logic (the rest of which has been removed).

Not only is this dead code (value is always NULL), but it additionally
conflicts with plans to register the `am_cq_handle` as the Cq to
receive target-side completion events (for target-side reassembly of
AM Long header + payload).