aries: inst_id and notify encoding
aries: inst_id and notify encoding

This commit changes the encoding of the lower 32-bits of the "notify"
word and the upper 8-bits of the "InstId".  While this introduces
small changes to the wire format and code, there are no behavioral
changes in this commit.

An enum has been created for the upper bits of the Instance Id, used
currently only to distinguish arrival of a control message (sent via
CqWrite) from a normal Put arrival.  The enum introduces two new cases
to cover the planned arrival of Long payloads (with Request and Reply
being distinct).

The "target_slot" moves into the upper most 6 bits (of the lower 32)
which had previously been unused bits in the "type" field (which needs
only 2).  The "initiator_slot" moves to the least-significant bits, so
that no shift is required to access it.

With these changes, bits 24-31 of the notify word and the InstId
are aligned such that they can each (with no additional manipulation)
become the corresponding bits of the "nonce" to be used for
target-side reassembly of Long header+payload.