aries: initial gasnetc_rdma_put_long()
This commit contains an initial gasnetc_rdma_put_long(), to provide
for non-trivial differences in the injection of AM Long payloads, as
compared to non-blocking Puts (the code for which was previously used
for Long payloads).

This makes several initial steps toward implementation of target-side
reassembly of Longs.  For instance there is logic (incomplete and
commented out) for setting the InstId to carry the necessary nonce;
and care is taken to ensure exactly one "final" xfer will generate
both a local event and (eventually) a remote event.  This latter
requires some more complex handling for the (unlikely) case that
multiple bounce-buffers are needed to complete the transfer (as can
now be tested using the recently-added testreadonly).  This is the
most significant departure from the non-blocking Put code path it is
based upon.

Some things are intentionally deferred to minimize changes to the
injection path (keeping these changes testable).  In particular the
code does not yet provide for any "nonce" and is still providing
semantics suitable for the pre-existing "put-sync-send" approach to AM
Longs.  Notably this second point means there is only a remote
completion indication, not yet any for LC.  Similarly, there is still
"IMMEDIATE" support in the payload injection, since it always precedes
the injection of the header.

Other than changes to name and signature of the function used to
perform the payload Put, the most significant changes to the injection
code are the removal of the "chunking" for the unlikely multiple
bounce-buffer case and the corresponding change to poll a boolean,
rather than counted, condition for completion of the payload transfer.