aries: begin generating payload Put notifications
This commit makes small modifications to the AM Long code paths to
begin passing a controlled "nonce" value in the Instance Id appearing
in the newly-generated target-side GNI-level remote event.  These are
generated when an RDMA or FMA transfer of a Long payload completes at
the target.

Currently the values used are simple 6-bit sequence numbers - one
each for Requests and Replies.  The only "consumer" is tracing (to be
removed).  The use of a sequential counter is not sufficient for the
final code, since it is subject to "aliasing" in the presence of
out-of-order retirement.  This is expected to be common when multiple
threads are polling and running handlers.  Additionally, no effort is
made to advance the counters atomically in a PAR build.  The use of
the sequence number will be replaced with the 6-bit "target slot"
value in the final implementation.

Since the "nonce" values are not sufficiently unique, this is just an
intermediate stage in the implementation, sufficient to demonstrate
the plumbing.  However, performance data collected shows that any
increase in AM Long latency is small relative to the normal level of
noise (though NOT zero).  The sources of the small increase in latency
include setting the InstId in the initiator's critical path; and the
generation and processing of a Cq event at the target.  It is expected
that overlap between header and payload transfers in the final
implementation will more than offset these added latencies.