aries: initial hash table for Long reassembly
This commit contains an initial implementation of a simple
chained-bucket hash table to be used for matching of an AM Long's
header and payload.

The hash table is (trivially) protected by an existing critical
section using the `ampoll_lock`.  Handing of the GNI-level events for
payload and header arrival are performed while holding that lock. In
the non-default RVous protocol this is the remote event for the notify
word, not the local event for the Get completion.

The arrival of either a header or a payload will check the hash table
for an existing entry.  If one is found (indicating this is the second
of the two arrivals) is it removed and (in the near future) either the
handler is run (Eager protocol) or the Get of the full header is
launched (RVous protocol).  Otherwise, a new hash table entry is
created at the tail of the bucket chain (assumes the normal order of
matching is much closer to FIFO than to LIFO).  The calls to the hash
operations are made in the proper locations, but as noted below, they
are not yet used to control the timing of handler execution.  However,
tracing (to be removed) is sufficient to verify the operation is as
intended.

The current implementation is still using a bogus 6-bit sequence
number in the InstId of payload Put events.  Therefore, this
implementation is ignoring the matching logic and continues the
pre-existing practice of running the handler upon arrival of the
header.  This implementation is correct because the injection is still
"put-sync-send".