aries: target-side reassembly of AM Longs
This commit provides the first complete implementation of target-side
reassembly of AM Long payload and header.  Specifically, the execution
of AM handlers now occurs at the latter of header arrival or payload
arrival.  Several things have changed in this commit, but they were
not practical to test independently.  The largest are described below.

The most important change, relative to previous versions of the code,
is that the payload Puts now carry the correct "nonce" value to permit
matching at the target.

The top-level logic for Long injection has changed to perform payload
injection after the header injection.  This is motivated by the fact
that the nonce value carried by the payload is among the resources
allocated to perform the header injection (specifically the "target
slot", naming the buffer at the target to which the header will be Put
in the Eager implementation).  Any approach that would allow us to
continue injecting payload before header would come with significantly
more complex logic to avoid deadlock (which could arise either from
holding multiple GPD's as in bug 2532, or by loss of atomicity in the
allocation of the multiple header-injection resources).

Because the header is now the first injection, there is no longer any
possibility of supporting IMMEDIATE injection of the payload Put.
This makes for small updates to both the top-level and Long injection
code and to gasnetc_rdma_put_long().  Note that this change is
probably a net *improvement* to the IMMEDIATE behavior, because the
injection of the payload first left no way to "fail" in the presence
of credit back pressure.  This is an argument against investing time
to pursue the efforts alluded to in the previous paragraph to resume
injection of (potentially large) payload prior to (consistently small)
header.

Smaller, but notable, changes include:

  + some factorization of post-Reply logic
  + simplification of the header-arrival logic
  + initial work toward having gasnetc_rdma_put_long() signal LC

While this implementation is now correct and functional complete,
there are at least three outstanding issues:

  + Complete work in gasnetc_rdma_put_long() and the main progress
    engine to allow Long injection to return following local (rather
    than remote) completion of the payload Put.
  + Follow up on the previous work to implement asynchronous LC
    (address-of-event and GEX_EVENT_GROUP).
  + Implement a GASNET_LONG_DEPTH environment variable to limit the
    growth of the Cq size.  Currently the size has doubled in both
    the Eager and RVous protocols.

Benchmarking using testam shows no difference in Short or Medium
sub-tests.  The Long ping-pong tests show a noticeable increase in
bandwidth at small payload size (4K being the smallest tested size not
using the "packed long" protocol).  This is attributed to the lack of
a stall between the payload and header injection.  Currently, the Long
flood bandwidth sub-tests show no measurable difference.