aries: allocate target slots for RVous
This commit modifies the AM flow control logic for the RVous protocol
to allocated a named "target slot" rather than just an anonymous
credit.  This means changing a counter into a bitmap, just as used in
the Eager protocol, except that the allocations are always single-bit.

This does not change the flow-control behavior (neither raising nor
lowering the number of credits) - just the encoding.  This change is
in preparation for target-side reassembly of Long payloads, in which
the [target_slot:jobrank] will be used as a unique identifier for
matching.