aries: env var GASNET_GNI_AMPOLL_BURST
This commit implements and documents a new environment variable
`GASNET_GNI_AMPOLL_BURST` which replaces the compiled-in `AM_BURST` as
the limit on the number of GNI-level events processed per Poll (while
retaining the default of 20).  The implementation uses `alloca()`
under normal circumstances, but is cautious to include an alternative
implementation (enabled for PGI) that avoids use of `alloca()` at the
expense of imposing a maximum value.

Additionally, this commit provides a new TRACE/STATS counter to track
how many GNI-level events are actually processed.

Finally, an unused `#include <alloca.h>` has been removed from
gasnet_core.c, where it was once used for loopback when PSHM was
disabled.  However, that use was eliminated when AM loopback support
was hoisted to conduit-independent code.