tests: improve coverage via comms layout
As observed in PR#288, many of the GASNet tests have historically used
`peer=(myrank^1)` to select a communication peer.  This has the risk
of exclusively utilizing the shared-memory communication channel in
any run with an even number of ranks per host.

This commit replaces the majority of those cases with a "circular"
assignment: `peer = (myrank + 1) % nranks`.  For tests which permit
runs with an odd number of ranks, the new logic preserves the "last
talks to self" mapping.

In this commit, all tests known to be using the xor-1 assignment have
been updated except for two tests: testam and testcoll.

+ Testam includes a ReqReq pattern in multiple sub-tests which
  requires strict pair-wise mappings.
+ Testcoll is focused on collective communication, but needs an RMA
  peer for testing for races against the sync flags.  In this case
  the tendency toward shared-memory communication may actually be
  resulting in a stronger test.  So, given the complexity of the
  change required, none has been made.

It is worth noting that testcore[245] have Request/Reply patterns
which required minor changes to accommodate sending Requests to one
rank while receiving Requests from a different one