ibv: create/allow intra-nbrhd connections
This commit changes connection and injection logic in ibv-conduit
to permit intra-neighborhood communication over IBV.  This is required
for multi-EP support, because non-primordial EPs are not cross-mapped by
PSHM.  Additionally, this will be a requirement for eventual off-loaded
remote atomics, which must perform all accesses via the HCA.  It is the
future case for atomics which motivates not retaining logic to exclude
loopback.

To compensate for the loss of "!IN_NBRHD" assertions at the deepest
levels of the call stack (where endpoint index info is unavailable),
this commit introduces new ones in the Put, Get and AM code paths.  In
the case of AMs the assertions is still "!IN_NBRHD".  However, for RMA
operations the assertion is "!PRIMORDIAL || !IN_NBDHB".