cross-configure-cray-aries: Make mpi-conduit optional
The cross-configure-cray-aries* cross-configure scripts previously
defaulted to forcing mpi-conduit. However this conduit is almost
never the best conduit to use on this system, and in some environments
may be unavailable due to module dependencies or other issues.

The new default probes for MPI support and only includes mpi-conduit
if it appears to be functional. The old behavior of force-enable can be
restored by passing `--enable-mpi` or alternatively `--disable-mpi`
can be used to force-disable the conduit.