fixup: aries: implement async LC for AM Longs
This commit corrects the logic for async local completion of Long
payloads.  Prior to this commit, an FMA-based transfer of an AM Long
payload would incorrectly increment the Put completion counter rather
than the ALC counter, because the FMA case generates only a GLOBAL
event and not a LOCAL one.  The fix is to consider eop- and iop-based
completion of Long payload transfers as distinct post-completion
actions as opposed to conflating them with Puts.

This issue went undetected in testing of the original code because by
default the Packed Long limit is about 4x the fma-to-rdma cutover,
rendering the FMA-based path unreachable.  However, in a tester
configured using `--with-aries-max-medium=512` the limit on Packed
Long falls accordingly and this code path became reachable.