ibv: resolve bug 4002 (odp/xrc conflict)
This commit adds logic to `gasneti_odp_init()` to query the XRC
support present in the ODP implementation (rather than RC), when
appropriate.  This will result in giving XRC precedence over ODP when
both are enabled but the underlying implementation cannot support them
together (as appears to be true of Mellanox's current stack).

Additionally, the warning generated for the case that XRC conflict is
the only reason for disabling ODP is updated accordingly.