ibv: cleanup support for two create_qp variants
This commit adds configure logic to detect availability of
`ibv_create_qp_ex()` rather than assuming it is present when the OFED
APIs for XRC are detected.

Additionally, the commit factors all instances of a create_qp call
(except the UD case for dynamic connections) and adds a `typedef`
for the corresponding attributes structure.

Finally, the XRC-related aspects of QP creation have been simplified
and some dead code removed.