ibv: fix regression with ODP and XRC
This commit corrects a regression introduced in PR#295.  The intent of
that PR was to automatically disable ODP when XRC was active.
However, the actual result was to disable ODP whenever XRC was
*available*, regardless of whether it was to be *used*.

This commit reorders the call to `gasneti_odp_init()` to come after
the one to `gasnetc_sndrcv_limits()`, where use of SRQ and XRC are
decided.  The ODP-related logic is not modified - just moved later in
the sequence of initialization steps.