Remove support for IBM XL compilers for big-endian
Historically we've supported IBM's XL compilers on big-endian PPC
(with their unique inline assembly syntax) on AIX, macOS and Linux.
Our support for AIX was abandoned some time ago, and Apple stopped
supporting PPC even before that.  That left only Linux on big-endian
PPC64.  With our retirement of pami-conduit and BG/Q support we no
longer have any target HPC platforms with this compiler, and we expect
to lose access for testing and maintenance at the end of 2019.

Builds with CC=xlc and CXX=xlC (on a BG/Q front-end) still compile and
pass smp-conduit PAR tests.  The gcc "sync" atomics are used, while
timers use GCC-style inline assembly (support for which is sufficient
for reading the timestamp register, but not for atomics).  In other
words: atomic, membars and timers remain "native" even after the
removal of the XL-specific assembly code.  Use of the "sync" atomics
is assumed to mildly degrade performance, since they are over-fenced.

It is worth noting that IBM withdrew support for big-endian PPC (XL
compilers or otherwise) from Open MPI in late 2017.

Note that XL compilers for little-endian CPUs are still supported.
They never contained support the `mcfunc` inline assembly which this
commit removes.