UPCXX CI: workaround issue278 - compile failures on issue138.cpp
test/regression/issue138.cpp constructs a deep chain of template
instantiations, which generates a combinatorial explosion of code in the
object file, especially when optimizations that reduce text size are
disabled.

A new `-DMINIMAL` preprocessor option activates a version of the test
that elides some of the types tested, pruning the combinatorial space
and trading test coverage for code size (and the ability to compile at
all on affected platforms).

This problem is particularly pronounced on Cygwin, where even the
MINIMAL version of the test reliably causes a fatal assembler error in
debug mode. As documented here:

https://stackoverflow.com/questions/16596876/object-file-has-too-many-sections

there is a problem in Windows' PE/COFF object format that prevents the
assembler from generating over 32k sections in one object file (the
original variant of the test generates over 130k sections). The addition
of `-g0 -Wa,-mbig-obj` in CI enables a workaround for the section limit
on windows, and also disables debug symbol output (which is also
enormous, causing a different assembler overflow).

Makefile needed a patch to add an override to the default of using
nobs to compile this test, because nobs is not amenable to adding compiler
options.