UPCXX CI: don't build issue138 on MIPS
This commit adds a ProhibitFeature annotation to the issue138 tests to
prevent building them on MIPS systems (both LE and BE).

This test code (even with `-DMINIMAL`) appears to exceed some ABI
limitations.  All the potential work-arounds I found online either
failed in my trials, or require one to rebuild libc and libgcc with an
alternative ABI.  So, given that MIPS is not a supported UPC++
platform, I have chosen to disable the tests entirely on this platform
over allowing them it burn significant memory and time generating a .o
that cannot be linked.