aries: correct rare leak of Long credits
This commit corrects a logic error which could have led to a "leak"
of the "Long credits" (necessary to send Long payloads), when
injecting an AM Long with GEX_FLAG_IMMEDIATE.  If the operation
returned without injection due to either an insufficient number of
reply buffers or post descriptors, then the Long credit already
allocated was not released.

With default settings, both the reply buffers and post descriptors are
more numerous than the other resources which must be obtained earlier.
As a result, it was very unlikely (potentially impossible?) for this
problem to occur in practice.  However, unusually small values of
`GASNET_NETWORKDEPTH_TOTAL` or `GASNET_GNI_NUM_PD` could have led to
this leak.