gasneti_nodemapInit: implement "Case 4"
This commit provides the previously unimplemented "Case 4" calling
convention for gasneti_nodemapInit().  Specifically, an exchangefn
and a local ID are accepted and the required exchange is performed
by the common code.  Essentially, this is a simple means for the
conduit to provide its own addressing information as a replacement
for the (not always reliable) gethostid().