UPCXX CI: Fix compiler settings for new cmake test
We need to provide cmake with the CC/CXX used by UPC++, overriding the
system default values which are often the wrong compiler.

Isolate all these envvars in a subshell around cmake, to ensure
the make step is not depending on them.

Compilers are passed to CMake via `$CC`/`$CXX` because -DCMAKE_{C,CXX}_COMPILER
does not support embedded compiler options, but the former does.