ibv: implement GASNET_AM_GATHER_MIN
This commit adds and documents GASNET_AM_GATHER_MIN.  This provides
for a minimum payload size at which multiple scatter-gather segments
(gather-on-send) are used to combine header and payload of a Medium or
packed Long (assuming all other requirements are met).  Previously the
size of the payload was not considered when choosing whether to use
gather-on-send.  Below this value, memcpy() is used to pack into a
single buffer.

This parameter and its default of 1500 bytes are based on the
observation that the overhead of gather-on-send is comparable or
greater than the cost of memcpy() up to approximately 1500 bytes on
systems with modern hardware.