ibv: use gather-on-send via ODP when available
This commit expands the set of circumstances under which
gather-on-send may be used (to combine the header and payload of an AM
Medium or packed Long) to include use of ODP, where previously only
in-segment payloads were considered.

Measurements on modern hardware show this provides the (small but
measurable) benefit seen for in-segment payloads when applied to
out-of-segment payloads.