Fix install problem with automake 1.4
make install after a Bootstrap using automake 1.4 has been mistakenly
omitting headers extended-ref/*/*.h since we rearranged those headers
in Apr 2018. This was causing a failure in the new tests-installed
target (which is designed to detect exactly such defects) under the
distcheck_old tester.

Automake 1.4 is not officially supported, so this is not a user-facing
issue.

This workaround is effective because AM 1.4's earlier attempt to install
without stripping the header prefix generates an error which is ignored.