fixup: mpi-spawner: replace `memcpy` with `GASN...
This commit corrects a regression due to an incorrect assumption that
GASNETI_MEMCPY() is an expression returning the destination.  This was
not noticed in initial testing because the problematic code is only
reachable for old MPI versions prior to 2.1 (lacking MPI_IN_PLACE
support in MPI_Alltoall and/or MPI_Allgather).