UPCXX CI: repair UPCXX_TARBALL_GASNET
This commit repairs the logic for UPCXX_TARBALL_GASNET with the new
infrastructure.  In particular the new infrastructure requires that
for `GASNET=foo.tar.gz` the tarball will contain a top-level directory
with the matching name `foo`.

This change also has two added side-effect.  This first is to avoid an
extra copy of the distribution tarball (which, unlike with nobs, is
both created and consumed in the span of the `_upcxx` test step).  The
second is that if more than one `GASNet-*.tar.gz` exists in the build
directory, the newest (just built) will be used.