Implement configure probe for GASNETI_COMMON
This commit adds a probe for the C compiler's behavior with respect to
tentative definitions.  It they "just work" as we have always
expected, then `GASNETI_COMMON` is empty.  However, if we observe the
sort of behavior seen with the current gcc-10 snapshot, then we
proceed to probe to determined if `__attribute__((__common__))` is
sufficient to resolve the problem (defining `GASNETI_COMMON` to that
attribute decoration if so).