amx: fix tentative definitions for GCC-10
This commit applies a rough analogue of the new `GASNETI_COMMON`
decorator on the uses of tentative definitions (of function pointers
in libgasnet) used within AMX to conditionally access debugging malloc
capabilities.

Unlike the case with libgasnet-*, when compiling libam{udp,mpi} we do
not have a configure script nor (at least currently) the ability to
leverage `gasnet_config.h`.  So, this commit is hard-coding the use of
the literal `__attribute__((__comon__))` and its use for GCC-10 and
newer.