ibv: implement gasnetc_rdma_long_put()
This commit creates a new `gasnetc_rdma_long_put()` to handle the
transfer of Long payloads, distinct from the path used for Put in the
Extended API.  This has multiple benefits.

1) The transfer of a Long payload is the only case in which a payload
must follow a specific path (named by a "qpi" == "queue pair index").
Prior to this commit a qpi was passed the RDMA Put and Get functions,
but now only to the new long-payload-specific put.

2) The injection of an AM Long payload always needs local completion
indication, but never needs remote completion indication.  This
removes arguments and branches from `gasnetc_rdma_long_put()` relative
to `gasnetc_rdma_put()`.  Additionally, a branch for the inline-send
case is removed, since anything that small would be sent as a Packed
Long.  Where `gasnetc_rdma_put()` for SEGMENT_LARGE has three
top-level cases (roughly inline, bulk and non-bulk), the new
`gasnetc_rdma_long_put()` has only one.  So, the volume of duplicated
code is fairly small.

3) In SEGMENT_EVERYTHING mode, remote firehose misses can require an
AM round-trip to resolve prior to actually injecting the Put.  ONLY
for the case of AM Long payload, we need to stall for this to complete
in order to ensure the header follows after the payload.  However,
prior to this commit there was an extra argument for *all* Puts to
support the caller performing the stall.  Since this argument was only
present in segment-everything mode, there was an asymmetry between
EVERYTHING and LARGE modes.  With this commit, the stall is now done
*within* `gasnetc_rdma_long_put()` (not its caller) and there is no
longer a need for a corresponding argument (and symmetry is restored).