ibv: create conduit-specific gasnetc_Segment_t
This commit creates an ibv-specific gasnetc_Segment_t, which is an
extension (as we intended in the design) of the top-level segment
object.  No use is made of the new object in this commit, which only
adds a single field used for management.  Real use will be in another
commit.