ibv: revisions for multi-segment support
This commit makes several inter-related changes to ibv-conduit's RMA
and AM communication paths to move toward multi-EP and multi-segment
support.

This commit relocates the registration keys (one per HCA) for the
local segment out of the per-HCA and per-QP data structures into the
(conduit-specific extension to the) Segment data structure, which is
then accessed from the Ep, and NOT from global variables.

This work is believed to be complete with respect to both the
conduit's implementation of the Extended API (Put and Get) and Core
API (both FPAM and NPAM) which should now always use the segment bound
to the caller's Ep (the one associated with the `tm` argument).  Here
"use" means that local in-segment bounds checks and registration keys
no longer use global data structures.  However, there is no means yet
to test this.

Note, however, that anywhere the Coll or VIS reference implementations
are using gasneti_THUNK_TM the true Ep and segment information has
been lost before reaching the conduit.