ibv: remove (debug only) timestamp from AM header
This commit removes a timestamp field from the AM header of debug
builds.  This was originally intended to measure the round-trip time
of Request-Reply traffic in traces.  However, as soon as credit
coalescing was added, that became less useful (timing at most half
the round trips).

While it is true that even the conduit developer never found much use
for this timestamp, the real reason for its removal is to eliminate a
difference in the header size for DEBUG and NDEBUG builds, thus
simplifying the documentation aspects of providing a configure option
to control the Max/LUB Medium.

Additionally, removal of the timestamp eliminates the need for the
`token` argument to `gasnetc_am_commit()` and has a cascading effect
on the chain of callers.