ibv: revisions for multi-ep support
This commit eliminates the global variable containing the
conduit-specific "cep" (connected end-point) table in favor of
a per-EP table.  This will be critical when there are multiple EPs.

The bulk of the changes are to pass the EP down through the call stack
to the point at which it is needed.  In the case of SEGMENT_EVERYTHING
this included stashing it with other data passed to the callback.

The connection-establishment logic (both static and dynamic) has only
been partially updated so far.  Update for dynamic-connection in the
presence of multi-EP will require wire-protocol changes that have yet
to be designed.