gasneti_in_auxsegment() takes a jobrank
This commit updates `gasneti_in_auxsegment()` to take only a jobrank,
rather than a `(tm,rank)` pair.  This is in recognition of the plan to
have a single aux segment per client, and not one per endpoint.  The
client argument will be added later when that becomes relevant.

This commit also splits `_gasneti_seginfo_lookup()` into separate
functions for client and aux segments.  In the process, the last
caller to `_gasneti_e_tm_rank_to_jobrank_allownull()` is removed
and that function is removed as well.