Add work-around for testcore2 on ibv+ILP32
Because testcore2 tests NPAM Large, with its malloc/free of temporary
payload buffers, it is susceptible to both firehose and ODP
misbehaviors with respect to unmapping of dynamically registered
memory.

While we have a recommended work-around, it is not effective for this
test on ILP32 targets.  This is most likely because the test's memory
usage is such that malloc cannot satisfy all requests from the sbrk
heap alone.

So, this commit limits testing on IBV+ILP32 to only AM Mediums
(documented as an effective work-around in bug 3989, comment 25).