amudp: startup warning for misconfigured DNS
Add detection for the case where a slave is given a master hostname
string that resolves to the localhost network, and generate a
warning if we subsequenetly fail to connect to that master.

This is intended to warn about systems with misconfigured DNS that
something "fishy" is going on before we issue a fatal
error after failing to connect to the master on the wrong network.

This failure mode could result from DNS on a slave node mis-resolving
a real master hostname to the localhost network, or could also arise
in a case where a misconfigured master believes/reports its own hostname
to be "localhost"