fixup: Tentative definitions overhaul
The overhaul changed the details of how gasnett_backtrace_user
was declared when compiling gasnet_tools.o for libgasnet, where
the common block definition is created for this tentative definition.
Specifically it removed a preceding extern declaration (in tools.h)
for that particular TU.

This shouldn't matter, because a tentative definition is implicitly
extern (by virtue of being non-static). However this seemingly
harmless change was sufficient to tickle a known bug in gcc LTO:
 https://sourceware.org/bugzilla/show_bug.cgi?id=23350
 https://gcc.gnu.org/bugzilla/show_bug.cgi?id=86490
that appears when -lgasnet-* appears twice on the link line, as is
routine when linking a UPC/UPC++ hybrid application.
The failure mode looked like this:
```
lto1: fatal error: multiple prevailing defs for
'gasneti_diagnostic_gethandlers'
compilation terminated.
lto-wrapper: fatal error: /usr/local/pkg/gcc/9.2.0/bin/g++ returned 1
exit status
compilation terminated.
/usr/bin/ld: error: lto-wrapper failed
```

This commit is a surgical reversion of that non-meaningful change,
ensuring LTO sees and processes this tentative defn as an extern symbol.
It repairs a regression affecting all hybrid UPC/UPC++ app links (and
possibly other usage scenarios) when LTO is enabled, which was only
discovered after the code freeze.