configure: use 'default' libdir for Cray libraries
This commit adds logic to the configure script to rewrite Cray's
CRAY_[LIB]_POST_LINK_OPTS environment variables to use the `default`
subdir (a symlink) in place of the fully versioned values.  This
ensures a degree of resilience in the presence of updates to `pmi`,
`ugni`, `udreg` and `xpmem` libraries.