aries: repair regression in startup time
This commit resolves a regression seen in the time required for job
start-up at scale, which is due to a configure-time failure to detect
availability of the Cray-specific `PMI_Allgather()` extension to PMI.

Evidence is consistent with this regression having been caused by a
change in the Cray `cc` wrapper such that `libpmi` is no longer linked
by default.  In particular, logs from nightly testing show the
regression in configure detection has been observable on Cori since
the return from maintenance on December 6, 2019.  Meanwhile the
problem is not occurring on Theta even today.

The nature of the commit is to restructure the pmi-spawner support in
aries-conduit to follow the pattern of all other conduits with such
support.  Specifically, the same `PMI_SPAWNER_{CFLAGS,LDFLAGS,LIBS}`
variables are used, rather then placing the PMI-specific elements into
the `GNI_*` equivalents.  Of course this is a "good thing" from a
general software engineering perspective.  More importantly, however,
this change ensures the probe for `PMI_{Allgather,Bcast}` will succeed
independent of which PMI-specific options (if any) the `cc` wrapper may
(or may not) inject.

Additionally, the `--with-pmi-{home,libs,ldflags}` (and equivalent
environment variables) can now be used to select an alternate Cray PMI
installation or adapt to unforeseen changes w/o source modification.

Tested on both Cori and Theta.  As noted above, these two Cray XC
systems display different behaviors in the absence of this commit.
With this commit, however, they both display the desired behavior.

This commit resolves bug 4061.