pmi-spawner: validate libpmix API version
This commit adds a check for new-enough PMIx library.  We support
(at least) versions 2 and 3 of the PMIx API, but not version 1.

If the user has passed `--with-pmi-version=x`, then detection of a
version 1 API implementation is fatal:
```
checking for PMIx API version 2 or newer... no
configure: WARNING: Ignoring libpmix implementing API version 1 because version 2 or newer is required.
checking for PMI spawner support... no

configure error: Requested PMI version x could not be found
```

If no specific version was requested, then the `WARNING:` above is
followed by probes for PMI2 (and PMI1 if necessary).  On platforms I
can test, the probe for PMI2 was successfully, since PMIx provides
`libpmi2` (and `libpmi`) for use by legacy clients.  So, in the end,
we use the same implementation but via a legacy API.

A portion of this commit is code movement to perform the check for
`PMI_SPAWNER_CFLAGS` individually with each of the three library
versions via new a new DEFUN: `GASNET_PMI_SPAWNER_CFLAGS`.  This is a
necessary in order to perform the preprocessor-based version check in
an order that allows falling-through to the probe for PMI2.