pmi-spawner: add garbage collection
This commit adds logic to the `Cleanup` function in pmi-spawner to
free (to the extent possible) the storage consumed in the KVS by
performing Puts of empty strings to all keys previous used (and ONLY
those keys).  Tracking of used keys is added to all the paths which
include `do_kvs_put()` calls.