pmi-spawner: render Cleanup non-final
This commit allows for the possibility of using pmi-spawner operations
after a call to `spawner->Cleanup()` by reallocating resources which
were freed if they are needed again.  This mainly includes lazy
allocation of the KVS buffers, but also makes changes to "GC" of the
KVS to ensure that `Cleanup()` can safely be called multiple times.

It is worth noting that pmi-spawner is the only standard spawner with
a non-empty `Cleanup()`, meaning they are all now "non-final".  So,
conduits can cleanup resources unconditionally at the end of Init and
still safely use them at Exit if needed.