pmi-spawner: no KVS space reduction by default
This commit introduces a preprocessor macro to conditionalize
two improvements related to KVS space reduction.  One is the
use of two "phases" of key (rather than counter based) to
bound the number of keys used (which also allowed for a
reduction in key length).  The second is "garbage collection",
consisting of overwriting (only) previously-used keys with
zero-length values (which also included logic to track key
use).

Testing has shown that of the five PMI *server* implementations
I have the ability to test, only SLURM's yields "nice" behavior
when overwriting previously used keys.  Despite the implication in
ANL's headers that multiple Puts from the same process separated
by a commit/fence should work, they don't (presumably because nobody
is testing that behavior).  The behaviors of the various servers
vary (and include abort at Put, failure at Get, and silently
out-of-date values from Get).

Due to intentionally compatible wire protocols, I do not have
any reliable means to determine the server in use.  For instance,
even if we compiled against the SLURM headers and can see
`$SLURM_JOB_ID` in the runtime environment, it is still possible
that ANLs `hydra` (aka `mpiexec`) was used rather than `srun`.
Thus the new behavior is to default to OFF for these optimizations
which depend on multiple Puts to the same key.